Moved MainListFilter from data to adt
